Update doctoshotgun.py
Builder pattern of creational design pattern was used to get different vaccines names and there associated keys for Germany and France sites. This is used to print out possible vaccines that is available to the country's citizens.

The builder pattern was selected among other creational design pattern in this case because it is known to isolate the building of a complicated item from its own representation, allowing the same construction process to generate several representations.

* feat 1: import from ABCMeta and abstractmethod to allow for abstraction
* feat 2: Builder pattern for vaccinations
* docs: Design Patterns in Python by Sean Bradley (https://medium.com/design-* patterns-in-python/builder-design-pattern-in-python-9a66de3e000b)
* Youtube res: https://www.youtube.com/watch?v=pMadA6F4zGU
* Lines: line 30 & line(605-678)