Update doctoshotgun.py
Added a simple skeletal of one of the structural design pattern, i.e. composite pattern, for the 1st and 2nd shots of vaccines for patients in France and/or Germany.

The idea behind using composite design pattern was to see how it enabled a client object to treat single components as wells as groups of components in the about the same way. How it adapted to the concepts of single responsibility by separating the aggregate of items form the characteristics of the projects.
*feat 1: import from ABCMeta and abstractmethod to allow for abstraction
*feat 2: structural design pattern for vaccinations
*docs: Composite Design Pattern (https://sbcode.net/python/composite/)
*Youtube res: https://www.youtube.com/watch?v=EWDmWbJ4wRA
*Lines: line 30 & line(605-659)