Creational Pattern Changes: Singleton
A singleton class is a design pattern that ensures that a class has at most one instance and provides global access to that instance. 
The getInstance( ) method implemented in the class AppointmentPage allows access to the singleton instance AppointmentPage_Instance. 
This implemented design allows for access into the AppointmentPage_Instance from a global scope where it is easy to use and implement. The implementation of the singleton AppointmentPage_Instance avoids bugs from a lot of many other instance that may be created therefore raising an exception when there is more than one instance of it.